fix: guard the MCMC spawn pool against notebook __main__ re-execution
The workers>1 MCMC pool started spawn workers directly, so a
`%run notebook.ipynb` session — where __main__.__file__ is the notebook
JSON — would crash every worker as spawn tries to runpy the JSON. The SbS
executor already guards this via sanitized_spawn_main; the MCMC pool did not.

Relocate sanitized_spawn_main from utils/sbs.py to a neutral utils/spawn.py
(fitlib importing sbs would be circular, and the helper is generic, not
SbS-specific) and wrap the MCMC pool with it, matching SbS.

Bump version to 0.13.1.
